Authorization data with JWT

+++ src/main/java/com/example/ProjectDAC/controller/CategoryController.java
@@ -11,12 +11,14 @@ import com.example.ProjectDAC.service.CategoryService;
 import com.example.ProjectDAC.service.UserService;
 import com.example.ProjectDAC.util.JwtUtils;
 import com.example.ProjectDAC.util.constant.ETypeCategory;
+import jakarta.persistence.Id;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -56,43 +58,72 @@ public class CategoryController {
     }
 
     @GetMapping("/category/{id}")
-    public ResponseEntity<Category> getCategory(@PathVariable("id") long id) throws IdInvalidException {
+    public ResponseEntity<Category> getCategory(@PathVariable("id") long id, HttpServletRequest request) throws IdInvalidException {
         Optional<Category> category = this.categoryService.getCategory(id);
         if(category.isEmpty()) {
             throw new IdInvalidException("Category does not exist");
         }
+        if(category.get().getAnken() == null) {
+            throw new IdInvalidException("You do not have permission to get this Category");
+        }
+        this.categoryService.checkPermission(request, category.get());
         return ResponseEntity.status(HttpStatus.OK).body(category.get());
     }
 
     @PutMapping("/category")
-    public ResponseEntity<Category> update(@RequestBody UpdateCategoryRequest request) throws IdInvalidException {
-        Optional<Category> categoryInDB = this.categoryService.getCategory(request.getId());
+    public ResponseEntity<Category> update(@RequestBody UpdateCategoryRequest updateRequest, HttpServletRequest request) throws IdInvalidException {
+        Optional<Category> categoryInDB = this.categoryService.getCategory(updateRequest.getId());
         if(categoryInDB.isEmpty()) {
             throw new IdInvalidException("Category does not exist");
         }
-        Category updatedCategory = this.categoryService.updateCategory(request, categoryInDB.get());
+        if(categoryInDB.get().getAnken() == null) {
+            throw new IdInvalidException("You do not have permission to get this Category");
+        }
+        this.categoryService.checkPermission(request, categoryInDB.get());
+        Category updatedCategory = this.categoryService.updateCategory(updateRequest, categoryInDB.get());
         return ResponseEntity.status(HttpStatus.OK).body(updatedCategory);
     }
 
     @DeleteMapping("/category/{id}")
-    public ResponseEntity<String> delete(@PathVariable("id") long id) throws IdInvalidException {
+    public ResponseEntity<String> delete(@PathVariable("id") long id, HttpServletRequest request) throws IdInvalidException {
         Optional<Category> categoryInDB = this.categoryService.getCategory(id);
         if(categoryInDB.isEmpty()) {
             throw new IdInvalidException("Category does not exist");
         }
+        if(categoryInDB.get().getAnken() == null) {
+            throw new IdInvalidException("You do not have permission to get this Category");
+        }
+        this.categoryService.checkPermission(request, categoryInDB.get());
         this.categoryService.delete(categoryInDB.get());
         return ResponseEntity.status(HttpStatus.OK).body("Delete Success");
     }
 
     @GetMapping("/category/account-category")
-    public ResponseEntity<List<ResCategoryDTO>> getCategoryAccount() {
+    public ResponseEntity<List<ResCategoryDTO>> getCategoryAccount(HttpServletRequest request) throws IdInvalidException {
+
         List<ResCategoryDTO> res = this.categoryService.getCategoriesByCategoryType(ETypeCategory.ACCOUNT);
-        return ResponseEntity.status(HttpStatus.OK).body(res);
+        List<ResCategoryDTO> result = new ArrayList<>();
+        String token = request.getHeader("Authorization");
+        List<Long> ids = this.jwtUtils.getAnkenListFromToken(token);
+        for(ResCategoryDTO resCategoryDTO : res) {
+            if(this.categoryService.checkIdIn(resCategoryDTO.getAnkenId(), ids)) {
+                result.add(resCategoryDTO);
+            }
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(result);
     }
 
     @GetMapping("/category/campaign-category")
-    public ResponseEntity<List<ResCategoryDTO>> getCategoryCampaign() {
+    public ResponseEntity<List<ResCategoryDTO>> getCategoryCampaign(HttpServletRequest request) throws IdInvalidException {
         List<ResCategoryDTO> res = this.categoryService.getCategoriesByCategoryType(ETypeCategory.CAMPAIGN);
-        return ResponseEntity.status(HttpStatus.OK).body(res);
+        List<ResCategoryDTO> result = new ArrayList<>();
+        String token = request.getHeader("Authorization");
+        List<Long> ids = this.jwtUtils.getAnkenListFromToken(token);
+        for(ResCategoryDTO resCategoryDTO : res) {
+            if(this.categoryService.checkIdIn(resCategoryDTO.getAnkenId(), ids)) {
+                result.add(resCategoryDTO);
+            }
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(result);
     }
 }
